moved, fixed
-moved all declared config properties to a separate header file
-made it so that validateConfigKeybinds is looping through keyBindNames and validating every single one
