+++ CMakeLists.txt
@@ -60,7 +60,7 @@ add_executable (
         "utilities/settingsMenu.cpp"
         "utilities/tesseract.cpp"
         "utilities/easyocr.cpp" 
-        "utilities/ocrfunctions.cpp" "utilities/settingsChanges.cpp" "utilities/settingsChanged.h" "utilities/relicDatabase.cpp")
+        "utilities/ocrfunctions.cpp" "utilities/settingsChanges.cpp" "utilities/settingsChanged.h" "utilities/relicDatabase.cpp" "utilities/configNames.h")
 
 
 target_link_libraries(${PROJECT_NAME} PRIVATE ${LIBRARIES})
