+++ utilities/settingsChanged.h
@@ -1,5 +1,4 @@
 #pragma once
-
 #include <vector>
 #include <string>
