+++ utilities/settingsChanges.cpp
@@ -1,18 +1,5 @@
 #include "settingsChanged.h"
-
-const std::vector<std::string> keyBindNames = {
-	"keyBind_ReadItemsFromScreen",
-	"keyBind_EscapeProgram",
-	"keyBind_ReadPreviousItems",
-	"keyBind_WindowVisibility",
-	"keyBind_BackupConfig",
-	"keyBind_ExampleItems",
-	"keyBind_ReadRelicTitle"
-};
-const std::vector<std::string> windowSizesNames = {
-"sfmlSize",
-"imguiSize"
-};
+#include "configNames.h"
